max-bot/requirements.txt
maxapi>=0.9.16
python-docx>=1.1.0
pydub>=0.25.1
httpx>=0.27.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
aiohttp>=3.9.0
boto3>=1.34.0
openpyxl>=3.1.0
